consolidate app components and add error boundary
