--- app/javascript/src/App.js
@@ -1,10 +1,0 @@
-// @flow
-import '@/styles/globals'
-import React from 'react'
-import { Layout, Routes } from './navigation'
-
-export const App = () => (
-  <Layout>
-    <Routes />
-  </Layout>
-)
