+++ app/javascript/src/App/Layout/Logo.js
