+++ app/javascript/src/App/Layout/MobileHeader/MobileHeader.js
