+++ app/javascript/src/App/Layout/MobileHeader/MobileMenuButton.js
