+++ app/javascript/src/App/Layout/MobileHeader/__tests__/MobileHeader.test.js
