+++ app/javascript/src/App/Layout/MobileHeader/index.js
