+++ app/javascript/src/App/Layout/__tests__/Layout.test.js
