+++ app/javascript/src/App/Routes/index.js
@@ -1,3 +1,2 @@
 // @flow
-export { Layout } from './Layout'
 export { Routes } from './Routes'
