+++ app/javascript/src/App/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { App } from './App'
